+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0 (unreleased)
 
   * Sanitize and upcase by default
+  * If `-reload` is not set, don't watch keys
 
 ## 0.1.0 (May 13, 2014)